DBZ-3528 Backport EventRouter refactors
